Script to uninstall cluster services
This script will remove the cluster services, allowing you to run
`cfy cluster start` (or join) again. This is useful only for development,
when you want to speed up working on the cluster configuration, or when
you passed the wrong flag (eg --cluster-host-ip) and want to try again.

DO NOT USE THIS IN PRODUCTION ENVIRONMENTS. This is NOT an official
"leave the cluster but stay active" script. Things might and will break,
including agents connected to the cluster.